% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_GR4J.R
\name{create_GR4J.data.frame}
\alias{create_GR4J.data.frame}
\title{create_GR4J}
\usage{
\method{create_GR4J}{data.frame}(d, inds_calib = NULL, inds_valid = NULL, perc_calib = 0.7, ...)
}
\arguments{
\item{d}{A data.table with the columns of \code{date}, \code{P}, \code{E}, \code{Robs}}
}
\description{
create_GR4J
}
