% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroALL.R
\docType{package}
\name{HydroALL-package}
\alias{HydroALL}
\alias{HydroALL-package}
\title{HydroALL: Hydrological Model Alliances in R}
\description{
Hydrological model alliances in R.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
